+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Update default indicator position in `SwitchControlItem`(Orange-OpenSource/ouds-ios#845)
 - Update `controlItem` component to use `controlItemBorderRadius*` tokens (Orange-OpenSource/ouds-ios#777)
 - Update `switch` component to use `switchBorderRadius*` tokens (Orange-OpenSource/ouds-ios#780)